Ajout de fonctionnalité (Liste projet) Customisation quasiment fini

+++ src/fileeditor.cpp
@@ -26,6 +26,7 @@
 #include "xmleditor.h"
 #include "jseditor.h"
 #include "xslproject.h"
+#include "xinxconfig.h"
 
 /* NumberBar */
 
@@ -443,7 +444,7 @@ bool FileEditor::saveFile( const QString & fileName ){
 		if( (!isOldSpecifiqueFile) && isNewSpecifiqueFile )
 			QFile::copy( m_fileName, destName );
 	}
-	if( fileName.isEmpty() || ( m_fileName == fileName )  ){
+	if( xinxConfig->isCreateBackupFile() && ( fileName.isEmpty() || ( m_fileName == fileName ) )  ){
 		if( QFile::exists( m_fileName + ".bak" ) ) 
 			QFile::remove( m_fileName + ".bak" );
 		QFile::copy( m_fileName, m_fileName + ".bak" );
